workers cannot work this way
Workers cannot host the main loop because they can handle just a task at a
time. Because of this, if the main loop is running without setInterval, the
worker will not respond to other messages (including "stop" requests).

What I was trying to do:

MAIN   |start         draw      draw  stop        "yay!"
             \       /         /          \       /
WORKER |------stepped---stepped------------stopped

What happens actually:

MAIN   |start        (draw)    (draw)  stop      "fuu-"
             \       /         /          x      /
WORKER |------stepped---stepped-----------stepped----

Actually, the "draw" messages never reach the main thread because the worker
doesn't release the CPU.

What I've learned: workers should be used in a strict request/response
fashion.  Running a task then sending a message that interrupts it is not
possible. If a task should be cancelable, it should be handled by a child
worker so that the parent remains responsive and capable of calling the
terminate API function on its child.