Major rewrite: new grid format
Now each cell is a number that stores both the staus (dead or alive) and the
number of neighbours (0...8).

This way, there's no need to access the neighbours of each cell on each step,
but just the neighbours of the cells that change status.

Physics module is gone, may be restored in the future.